--- src/components/composites/book/detail/record/data/index.ts
@@ -1,2 +1,0 @@
-export { default as BookRecordListData } from './BookRecordListData';
-export { default as BookRecordListDataSkeleton } from './BookRecordListData.skeleton';
