+++ src/components/composites/book/detail/record/index.ts
@@ -1,1 +1,1 @@
-export * from './data';
+export * from './list';
