+++ src/components/composites/book/detail/record/list/index.ts
@@ -1,0 +1,2 @@
+export { default as BookRecordList } from './BookRecordList';
+export { default as BookRecordListSkeleton } from './BookRecordList.skeleton';
